renamed ResultDataHelper moved methods from this class to ParserServiceImpl
